rxv enhancements (#22)
* multi zone support

Add zone support to RXV object. The number of zones supported by a
device is discovered with the desc.xml.

By default all actions occur with Main_Zone, however a
zone_controllers() method provides a list of RXV objects, one per
zone, that will directly control each zone separately.

* add a function that directly plays a netradio address

This simplifies calling the library to directly play a netradio
address.

* default desc url to be based on ctrl url

In almost all cases the desc url is really rooted on the same url as
the ctrl url, so it seems reasonable to default it to being based on
the same url. You can still manually override if you know the value.

* make play_status work with NET_RADIO

The play_status has different attributes depending on the source, when
on the NET_RADIO source there is no Artist field, and there is an
additional Station field. This changes the way the XML is processed so
we'll optionally set the fields that exist, and set everything else to
"".