Add initial automated testing
This adds a couple of basic tests which can run without any real
hardware using requests-mock to handle the interactions.

The previous live test_integration is moved to check_integration so
that it's not discovered and run during a tox run. It is still runable
as before manually to do live testing.

Tox.ini is updated to skip interpeters that don't exist, and do
coverage during every run.