Introduce logging for failed requests.
In response to Home Assistant bug -
https://github.com/home-assistant/home-assistant/issues/4226

When things go wrong with a request in unexpected ways, we really
don't have enough information to figure out what just happened. This
is especially true with hardware generations that none of the upstream
developers have.

This adds logging support to the library, and introduces a couple of
error cases where we will log low level details of a failure to help
debug these kinds of issues in the future.