Round to nearest 0.5 db when setting volume
The rxv library interface passes around well understood db volumes
over it's interface, values such as -53.5db and the like. The Yamaha
API operates in terms of deci-db as an integer - i.e. 535.

However, most (if not all) Yamaha receivers have a 5 deci db
resolution, and trying to set values such as 531 cause an error. This
change makes the rxv library round all set calls up to the nearest
half db to remove these errors. These are especially triggered when
coming from Home Assistant which has it's own 0 - 100 idea of volume,
so is very often not hitting a defined db boundary.

Closes Bug #35